FIX: Token PK 데이터 타입에 맞춰 수정

+++ src/main/java/com/aivle/bookapp/domain/Token.java
@@ -26,7 +26,7 @@ public class Token {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "token_id")
-    private Long tokenId;
+    private Integer tokenId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(
